♻️ Extract the common functions into a new single object.
