+++ gearmeshing_ai/mcp_client/strategy/direct.py
@@ -15,9 +15,10 @@ from gearmeshing_ai.mcp_client.schemas.core import (
     ToolCallResult,
     TransportType,
 )
+from .base import StrategyCommonMixin
 
 
-class DirectMcpStrategy:
+class DirectMcpStrategy(StrategyCommonMixin):
     """
     Strategy for directly connecting to MCP servers.
 
@@ -130,29 +131,4 @@ class DirectMcpStrategy:
             headers["Authorization"] = cfg.auth_token
         return headers
 
-    def _infer_arguments(self, input_schema: Dict[str, Any]) -> List[ToolArgument]:
-        args: List[ToolArgument] = []
-        # Very light inference: if schema has top-level properties, map to arguments
-        props = input_schema.get("properties") if isinstance(input_schema, dict) else None
-        required = set(input_schema.get("required") or []) if isinstance(input_schema, dict) else set()
-        if isinstance(props, dict):
-            for k, v in props.items():
-                if not isinstance(v, dict):
-                    continue
-                typ = v.get("type") if isinstance(v.get("type"), str) else "string"
-                desc = v.get("description") if isinstance(v.get("description"), str) else None
-                args.append(
-                    ToolArgument(
-                        name=str(k),
-                        type=str(typ),
-                        required=str(k) in required,
-                        description=desc,
-                    )
-                )
-        return args
-
-    @staticmethod
-    def _is_mutating_tool_name(name: str) -> bool:
-        n = name.lower()
-        prefixes = ("create", "update", "delete", "remove", "post_", "put_", "patch_", "write", "set_")
-        return n.startswith(prefixes)
+    # _infer_arguments and _is_mutating_tool_name inherited from StrategyCommonMixin
